:sparkles: add setup for api tests using DB
