+++ adminSiteServer/sql-fixtures.d.ts
@@ -1,0 +1,1 @@
+declare module "sql-fixtures"
