+++ jest.db.config.js
@@ -4,7 +4,10 @@ module.exports = {
         {
             displayName: { name: "db", color: "blue" },
             testEnvironment: "node",
-            testMatch: ["<rootDir>/itsJustJavascript/db/tests/**/*.test.js"],
+            testMatch: [
+                "<rootDir>/itsJustJavascript/db/tests/**/*.test.js",
+                "<rootDir>/itsJustJavascript/adminSiteServer/*.test.(jsx|js)",
+            ],
         },
     ],
 }
